requirements.txt
flask>=1.1.2
flask_cors>=4.0.1
gunicorn>=20.1.0
requests>=2.27.1
numpy>=1.14.0
Pillow>=5.2.0
opencv-python-headless>=4.5.5.64
deepface>=0.0.79
tensorflow>=2.10.0
tf-keras>=2.10.0